Move the functions related to PyPI and the requirements files to separate files.
