+++ scripts/plugin_runner/python_venv/common.py
@@ -1,104 +1,14 @@
-import json
 import logging
-import os
-import re
 import subprocess
-from collections import defaultdict
 from copy import copy
 from distutils.version import Version
 from pathlib import Path
-from typing import Dict, List, Optional, Set, Tuple
+from typing import Dict, Optional
 
-from pkg_resources import parse_requirements as parse_line, parse_version
-
-import requests
-from requests.adapters import HTTPAdapter
-
-from urllib3 import Retry
-
-from utils.file_utils import get_all_file_system_items
-
-PYPI_PACKAGE_METADATA_URL = 'https://pypi.org/pypi/{package_name}/json'
-REQUIREMENTS_FILE_NAME_REGEXP = r'^[\S]*requirements[\S]*\.txt$'
-
-Specs = Set[Tuple[str, Version]]
-Requirements = Dict[str, Specs]
+from utils.python.pypi_utils import check_package_exists, get_available_versions
+from utils.python.requirements_utils import Requirements
 
 logger = logging.getLogger(__name__)
-
-
-def _normalize_requirement_name(name: str) -> str:
-    """
-    Normalize the string: the name is converted to lowercase and all dots and underscores are replaced by hyphens.
-
-    :param name: Name of package
-    :return: Normalized name of package
-    """
-    normalized_name = name.lower()
-    normalized_name = normalized_name.replace('.', '-')
-    normalized_name = normalized_name.replace('_', '-')
-    return normalized_name
-
-
-def gather_requirements_file_paths(root: Path) -> List[Path]:
-    """
-    Gather paths to all files with requirements on the passed path.
-
-    :param root: The path to the folder where you want to find the requirements files.
-    :return: List of file paths.
-    """
-    requirements_file_paths = get_all_file_system_items(
-        root=root,
-        item_condition=lambda name: re.match(REQUIREMENTS_FILE_NAME_REGEXP, name) is not None,
-    )
-
-    # TODO: handle symlinks
-    requirements_file_paths = list(filter(lambda path: os.path.isfile(path), requirements_file_paths))
-
-    logger.info(f'{len(requirements_file_paths)} requirement files have been collected.')
-
-    return requirements_file_paths
-
-
-def gather_requirements(root: Path) -> Requirements:
-    """
-    Gather the requirements from the requirements files contained in the passed path.
-
-    :param root: The path to the folder that contain the requirements files.
-    :return: Dictionary, where for each package the collected specs are listed.
-             The spec is a pair of operator and version.
-    """
-    logger.info('Collecting requirements.')
-
-    requirements = defaultdict(set)
-    requirements_file_paths = gather_requirements_file_paths(root)
-
-    for file_path in requirements_file_paths:
-        file_requirements = []
-        with open(file_path, encoding='utf8', errors='ignore') as file:
-            for line in file.readlines():
-                try:
-                    file_requirements.extend(list(parse_line(line)))
-                except Exception:
-                    # For some reason you can't catch RequirementParseError
-                    # (or InvalidRequirement), so we catch Exception.
-                    logger.warning(f'Unable to parse line "{line.strip()}" in the file {str(file_path)}. Skipping.')
-                    continue
-
-        for requirement in file_requirements:
-            specs = {(operator, parse_version(version)) for operator, version in requirement.specs}
-            requirements[_normalize_requirement_name(requirement.key)] |= specs
-
-    logger.info(f'Collected {len(requirements)} packages.')
-
-    return requirements
-
-
-def _create_session() -> requests.Session:
-    session = requests.Session()
-    retries = Retry(total=10, backoff_factor=0.1)
-    session.mount('https://', HTTPAdapter(max_retries=retries))
-    return session
 
 
 def filter_unavailable_packages(requirements: Requirements) -> Requirements:
@@ -111,59 +21,18 @@ def filter_unavailable_packages(requirements: Requirements) -> Requirements:
     """
     logger.info('Filtering unavailable packages.')
 
-    session = _create_session()
     filtered_requirements = copy(requirements)
     for package_name in requirements.keys():
         logger.info(f'Checking {package_name}.')
 
-        try:
-            response = session.get(PYPI_PACKAGE_METADATA_URL.format(package_name=package_name))
+        if check_package_exists(package_name):
+            continue
 
-            if response.status_code == 200:
-                continue
-
-            if response.status_code == 404:
-                logger.warning(
-                    f'The {package_name} package does not exist on PyPI. Removing the package from requirements.',
-                )
-                filtered_requirements.pop(package_name)
-            else:
-                logger.warning(
-                    f'PyPI returned an unexpected code ({response.status_code}). Skipping.',
-                )
-
-        except requests.exceptions.RequestException:
-            logger.error('An error occurred when accessing the PyPI. Skipping.')
+        logger.warning(f'The {package_name} package does not exist on PyPI. Removing the package from requirements.')
+        filtered_requirements.pop(package_name)
 
     logger.info(f'Filtered {len(requirements) - len(filtered_requirements)} packages.')
     return filtered_requirements
-
-
-def _get_available_versions(package_name: str) -> Set[Version]:
-    """
-    By a given package, collect a list of all the versions available on PyPI.
-
-    :param package_name: PyPI package name.
-    :return: Set of available versions. If the version could not be obtained, None will be returned.
-    """
-    session = _create_session()
-    url = PYPI_PACKAGE_METADATA_URL.format(package_name=package_name)
-    try:
-        metadata = session.get(url).json()
-    except requests.exceptions.RequestException:
-        logger.error('An error occurred when accessing the PyPI. Skipping.')
-        return set()
-    except json.JSONDecodeError:
-        logger.error(f'Failed to get a version for the {package_name} package. Skipping')
-        return set()
-
-    if 'releases' in metadata.keys():
-        versions = metadata['releases'].keys()
-    else:
-        logger.error('The PyPI response does not contain the "releases" field. Skipping.')
-        return set()
-
-    return set(map(parse_version, versions))
 
 
 def filter_unavailable_versions(specs: Requirements) -> Requirements:
@@ -181,7 +50,7 @@ def filter_unavailable_versions(specs: Requirements) -> Requirements:
     for package_name, package_specs in specs.items():
         logger.info(f'Checking {package_name} versions.')
 
-        available_versions = _get_available_versions(package_name)
+        available_versions = get_available_versions(package_name)
 
         filtered_specs = package_specs
         if available_versions:
